Add functionality for uploading files
Previously we always were uploading files in the browser so our backends
didn't need to actually upload files. Now the Tessa gem will allow
uploading files, and there's a nice macro on the Tessa::Asset class that
will handle error cases and return the newly created asset if successful
and false if there was a failure.