Upload files properly
We were encoding the file, but the endpoint expects the raw file data
not a 64bit encoded version like a multipart form post. This was causing
corrupted data.

One issue with this approach is that large files are loaded into memory
before the request can be run. A streaming approach would be preferred.
We should look into using the http gem instead of Faraday. It would
simplify a few things and allow us to pass an IO object instead of a raw
string.